Struct wasmer_sys_utils::memory::fd_memory::memories::VMMemory

pub struct VMMemory(pub Box<dyn LinearMemory + 'static>);

Represents linear memory that can be either owned or shared

Tuple Fields
0: Box<dyn LinearMemory + 'static>
Implementations

impl VMMemory

pub fn new(memory: &MemoryType, style: &MemoryStyle) -> Result<Self, MemoryError>

Creates a new linear memory instance of the correct type with specified minimum and maximum number of wasm pages.

This creates a Memory with owned metadata: this can be used to create a memory that will be imported into Wasm modules.

pub fn get_runtime_size(&self) -> u32

Returns the number of pages in the allocated memory block

pub unsafe fn from_definition(
memory: &MemoryType,
style: &MemoryStyle,
vm_memory_location: NonNull<VMMemoryDefinition>
) -> Result<Self, MemoryError>

Create a new linear memory instance with specified minimum and maximum number of wasm pages.

This creates a Memory with metadata owned by a VM, pointed to by vm_memory_location: this can be used to create a local memory.

Safety

vm_memory_location must point to a valid location in VM memory.

pub fn from_custom<IntoVMMemory>(memory: IntoVMMemory) -> Selfwhere
IntoVMMemory: Into<Self>,

Creates VMMemory from a custom implementation - the following into implementations are natively supported

VMOwnedMemory -> VMMemory

Box<dyn LinearMemory + ’static> -> VMMemory

pub fn copy(&mut self) -> Result<Box<dyn LinearMemory + 'static>, MemoryError>

Copies this memory to a new memory