Struct wasmer_sys_utils::memory::fd_memory::VMSharedMemory

pub struct VMSharedMemory {
    mmap: Arc<RwLock<WasmMmap>>,
    config: VMMemoryConfig,
    conditions: ThreadConditions,
}

A shared linear memory instance.

Fields
mmap: Arc<RwLock<WasmMmap>>config: VMMemoryConfigconditions: ThreadConditions
Implementations

impl VMSharedMemory

pub fn new( memory: &MemoryType, style: &MemoryStyle ) -> Result<Self, MemoryError>

Create a new linear memory instance with specified minimum and maximum number of wasm pages.

This creates a Memory with owned metadata: this can be used to create a memory that will be imported into Wasm modules.

pub unsafe fn from_definition( memory: &MemoryType, style: &MemoryStyle, vm_memory_location: NonNull<VMMemoryDefinition> ) -> Result<Self, MemoryError>

Create a new linear memory instance with specified minimum and maximum number of wasm pages.

This creates a Memory with metadata owned by a VM, pointed to by vm_memory_location: this can be used to create a local memory.

Safety

vm_memory_location must point to a valid location in VM memory.

pub fn copy(&mut self) -> Result<Self, MemoryError>

Copies this memory to a new memory